Move InventoryGUINotification into src/, include/ files
