+++ src/inventory_gui/notification.cpp
@@ -1,0 +1,52 @@
+
+
+
+#include <inventory_gui/notification.hpp>
+
+#include <allegro_flare/gui/surface_areas/box.h>
+#include <global_constants.hpp>
+#include <style_assets.hpp>
+
+
+
+InventoryGUINotification::InventoryGUINotification(UIWidget *parent)
+   : UIWidget(parent, "InventoryGUINotification", new UISurfaceAreaBox(SCREEN_W/2, SCREEN_H/4*3, 900, 100))
+   , visibility_timer(-1)
+{
+}
+
+
+
+void InventoryGUINotification::on_timer()
+{
+   visibility_timer -= 1.0/60.0;
+   if (visibility_timer < 0) visibility_timer = -1.0;
+}
+
+
+
+void InventoryGUINotification::mouse_down_func()
+{
+   UIWidget::mouse_down_func();
+   visibility_timer = -1.0;
+}
+
+
+
+void InventoryGUINotification::on_draw()
+{
+   if (visibility_timer < 0) return;
+   Style::draw_button(Style::NORMAL, 0, 0, place.size.x, place.size.y, notification_text);
+   //Style::draw_text_box(0, 0, place.size.x, place.size.y, notification_text);
+}
+
+
+
+void InventoryGUINotification::show(std::string text)
+{
+   notification_text = text;
+   visibility_timer = 5.0;
+}
+
+
+
